add navigation between character list and details screens

+++ core/domain/src/main/java/com/myaxa/core/domain/CharacterId.kt
@@ -1,3 +1,3 @@
 package com.myaxa.core.domain
 
-typealias CharacterId = Long
+typealias CharacterId = Int
